Downloads Polity IV

Downloads the Polity IV data set. It keeps specified variables and creates a standard country ID variable that can be used for merging the data with other data sets.

Usage

PolityGet(
  url = "http://www.systemicpeace.org/inscr/p4v2015.sav",
  vars = NULL,
  OutCountryID = "iso2c",
  standardCountryName = TRUE,
  na.rm = TRUE,
  duplicates = "message",
  fromLast = FALSE
)

Arguments

url

character string. The URL for the Polity IV data set you would like to download. Note: it must be for the SPSS version of the file.

vars

character vector containing the variables to keep. If vars = NULL then the entire data set is returned. Note that the country and year variables are always returned.

OutCountryID

character string. The type of country ID you would like to include in the output file along with the country name. See countrycode for available options.

standardCountryName

logical. Whether or not to standardise the country names variable based on country.name from countrycode.

na.rm

logical. Drop observations where OutCountryID is NA.

duplicates

character specifying how to handle duplicated country-year observations. Can be set to none to do nothing, message to simply report duplicates, drop to report and drop duplicates, and return to return a data frame with only duplicated observations (see also fromLast).

fromLast

logical indicating if duplication should be considered from the reverse side. Only relevant if duplicates = 'drop' or duplicates = 'out'.

Value

a data frame

See also

Examples

if (FALSE) {
# Download full data set
PolityData <- PolityGet()

# Create data frame with only the main Polity democracy variable (polity2)
Polity2Data <- PolityGet(vars = 'polity2',
                         OutCountryID = 'imf')
}