Create Winset and Selectorate variables

Creates the winset (W) and a modified version of the selectorate (S) variable from Bueno de Mesquita et al. (2003) using the most recent data available from Polity IV and the Database of Political Institutions.

WinsetCreator(PolityUrl = "http://www.systemicpeace.org/inscr/p4v2015.sav",
  DpiUrl = "http://bit.ly/1jZ3nmM", OutCountryID = "iso2c", na.rm = TRUE)

Source

See Bueno de Mesquita Bruce, Alastair Smith, Randolph M. Siverson, and James D. Morrow. 2003. The Logic of Political Survival. Cambridge, MA: MIT Press.

Morrow, J. D., Bueno De Mesquita, B., Siverson, R. M., and Smith, A. 2008. Retesting Selectorate Theory: Separating the Effects of W from Other Elements of Democracy. American Political Science Review, 102(03), 393-400.

Polity IV: http://www.systemicpeace.org/.

Modified so that military regime is take from Database of Political Institutions (http://go.worldbank.org/2EAGGLRZ40), rather than Banks, Arthur S. 1996. Political Handbook of the World. New York: CSA Publications.

Arguments

PolityUrl

character string. The URL for the Polity IV data set you would like to download. Note: it must be for the SPSS version of the file.

DpiUrl

character string. The URL for the Database of Political Institutions data set you would like to download. Note: the link must be to a Stata formated file.

OutCountryID

character string. The type of country ID you would like to include in the output file along with the country name. See countrycode for available options.

na.rm

logical. Drop observations where OutCountryID is NA.

Value

Returns a data frame with the following columns:

iso2c: The ISO two letter country code identifying the country. This can be changed to another country ID system using OutCountryID

country: Country names.

year: The year.

W: The winset variable. The variable is the same as Bueno de Mesquita (2003) except the military executive component is from DPI, rather than Banks (1996).

ModS: The modified selectorate variable. Instead of being based on Polity LEGSELEC, which is no longer create, ModS is based on the Legislative Index of Electoral Competetiveness variable (LIEC) from DPI. No legislature = 1. Unelected legislature = 2. Elected legislature with only 1 candidate = 3. Elected legislatures with 1 party and multiple candidates = 4. All other legislatures = 5. The variable is then standardised between 0 and 1 by subtracting by 1 then dividing by 4. Note: some countries were coded in LIEC as 3.5, so ModS has 6 levels between 0 and 1.

See also

Examples

if (FALSE) {
# Create winset data using default options
WinsetData <- WinsetCreator()
}